[BE] finish notice

See merge request s07-webmobile1-sub2/TERAS!28

+++ backend/src/main/java/com/teras/db/repository/NoticeRepository.java
@@ -13,9 +13,9 @@ import com.teras.db.entity.ClassEntity;
 import com.teras.db.entity.Notice;
 
 @Repository
-public interface NoticeRepository extends JpaRepository<Notice, Integer> {
+public interface NoticeRepository extends JpaRepository<Notice, Long> {
 
-	Optional<Notice> findByNoticeNo(int noticeNo);
+	Optional<Notice> findByNoticeNo(long noticeNo);
 
 	Optional<List<Notice>> findAllByOrderByNoticeNoDesc();
